Introduce flakechecker library for usage by external components.
A new library with one external function is introduced:

func CheckIfFlakeOccurred{}

This function checks the input string, loads the json pattern(s)
from the default or given directory and returns the list of flakePatterns.

It also returns clear information if the retry should happen or not as not
all the flakes may require retry.

Signed-off-by: Michal Pryc <mpryc@redhat.com>